refactor: extract shared execute_command() into chibi-core (#143)
both chibi-cli and chibi-json now delegate command dispatch, lifecycle
(init/shutdown), and housekeeping (auto-destroy, context touch, cache
cleanup) to core's execute_command(). binaries retain only their own
concerns: CLI keeps session management, context name resolution,
markdown rendering, image cache; JSON keeps stdin parsing and trust
mode.

- new execution.rs with CommandEffect enum for binary-side reactions
- emit_markdown() on OutputSink for markdown-bearing inspect/log content
- CLI pre-resolves "-", "new", "new:prefix" names before calling core
- ~950 lines removed across both binaries
