+++ AGENTS.md
@@ -46,6 +46,7 @@ chibi-cli (binary)   chibi-json (binary)
 - `agents_md.rs` — AGENTS.md discovery and loading (VCS-aware hierarchy)
 - `vcs.rs` — VCS root detection (`.git`, `.hg`, etc.)
 - `index/` — Codebase indexing (SQLite WAL, symbol extraction, language plugin interface)
+- `execution.rs` — Shared command execution (`execute_command`, `CommandEffect`)
 - `input.rs` — Core input types (`Command`, `ExecutionFlags`, `Inspectable`)
 - `output.rs` — `OutputSink` trait (abstraction over CLI text / JSON output)
