+++ crates/chibi-cli/src/config.rs
@@ -276,6 +276,12 @@ pub struct ResolvedConfig {
 impl ResolvedConfig {
     /// Get a config field value by path.
     /// First checks presentation fields, then delegates to core.
+    ///
+    /// TODO: currently unused since inspect moved to core, which only sees
+    /// CoreResolvedConfig. CLI-specific fields (render_markdown, image.*, etc.)
+    /// are no longer inspectable via `-n config.<field>`. Re-wire when we add
+    /// a config inspection hook or move inspect back to binaries.
+    #[allow(dead_code)]
     pub fn get_field(&self, path: &str) -> Option<String> {
         match path {
             "render_markdown" => Some(self.render_markdown.to_string()),
