+++ crates/chibi-cli/src/main.rs
@@ -10,23 +10,16 @@ mod output;
 mod session;
 mod sink;
 
-use chibi_core::context::{
-    Context, ContextEntry, ENTRY_TYPE_MESSAGE, ENTRY_TYPE_TOOL_CALL, ENTRY_TYPE_TOOL_RESULT,
-    now_timestamp,
-};
-use chibi_core::input::{Command, DebugKey};
+use chibi_core::input::Command;
 
-use crate::cli::{Cli, InspectableExt};
+use crate::cli::Cli;
 use crate::config::{ImageConfig, ResolvedConfig, default_markdown_style, load_cli_config};
 use crate::input::{ChibiInput, ContextSelection, UsernameOverride};
 use crate::markdown::{MarkdownConfig, MarkdownStream};
 use crate::output::OutputHandler;
 use crate::session::Session;
 use crate::sink::CliResponseSink;
-use chibi_core::{
-    Chibi, Inspectable, LoadOptions, OutputSink, PermissionHandler, PromptOptions, StatePaths, api,
-    tools,
-};
+use chibi_core::{Chibi, CommandEffect, LoadOptions, OutputSink, PermissionHandler, StatePaths};
 use std::io::{self, ErrorKind, Write};
 use std::path::PathBuf;
 
@@ -184,262 +177,66 @@ fn resolve_cli_config(
     })
 }
 
-/// CLI display and behaviour flags for the send path.
+/// Resolve CLI-specific context names in Command variants.
 ///
-/// Groups the flags that are constant across all `send_with_cli_sink` calls
-/// within a single `execute_from_input` invocation.
-struct CliSendOptions<'a> {
-    raw: bool,
-    no_tool_calls: bool,
-    debug: &'a [DebugKey],
-    verbose: bool,
-    force_markdown: bool,
-    show_tool_calls: bool,
-    show_thinking: bool,
+/// Handles "-" (previous context), "new", and "new:prefix" in optional
+/// context name fields. Core doesn't know about these — they're CLI
+/// conveniences tied to session state.
+fn resolve_command_names(
+    command: &Command,
+    chibi: &Chibi,
+    session: &Session,
+) -> io::Result<Command> {
+    // Helper: resolve an optional name, defaulting to working_context
+    let resolve_opt = |name: &Option<String>| -> io::Result<Option<String>> {
+        match name {
+            Some(n) => Ok(Some(resolve_context_name(chibi, session, n)?)),
+            None => Ok(None),
+        }
+    };
+
+    match command {
+        Command::DestroyContext { name } => Ok(Command::DestroyContext {
+            name: resolve_opt(name)?,
+        }),
+        Command::ArchiveHistory { name } => Ok(Command::ArchiveHistory {
+            name: resolve_opt(name)?,
+        }),
+        Command::CompactContext { name } => Ok(Command::CompactContext {
+            name: resolve_opt(name)?,
+        }),
+        Command::RenameContext { old, new } => Ok(Command::RenameContext {
+            old: resolve_opt(old)?,
+            new: new.clone(),
+        }),
+        Command::ShowLog { context, count } => Ok(Command::ShowLog {
+            context: resolve_opt(context)?,
+            count: *count,
+        }),
+        Command::Inspect { context, thing } => Ok(Command::Inspect {
+            context: resolve_opt(context)?,
+            thing: thing.clone(),
+        }),
+        Command::SetSystemPrompt { context, prompt } => Ok(Command::SetSystemPrompt {
+            context: resolve_opt(context)?,
+            prompt: prompt.clone(),
+        }),
+        Command::ClearCache { name } => Ok(Command::ClearCache {
+            name: resolve_opt(name)?,
+        }),
+        Command::CheckInbox { context } => Ok(Command::CheckInbox {
+            context: resolve_context_name(chibi, session, context)?,
+        }),
+        // All other commands pass through unchanged
+        _ => Ok(command.clone()),
+    }
 }
 
-/// Resolve config, acquire lock, build sink, and send a prompt through the agentic loop.
+/// Execute from ChibiInput.
 ///
-/// This is the shared "send prompt" path used by SendPrompt, CallTool (with
-/// continuation), CheckInbox, and CheckAllInboxes. Keeps the resolve → lock →
-/// build-sink → stream sequence in one place.
-async fn send_with_cli_sink(
-    chibi: &Chibi,
-    ctx_name: &str,
-    prompt: &str,
-    ephemeral_username: Option<&str>,
-    opts: &CliSendOptions<'_>,
-    fallback: Option<chibi_core::tools::HandoffTarget>,
-    output: &dyn OutputSink,
-) -> io::Result<()> {
-    let mut resolved = resolve_cli_config(chibi, ctx_name, ephemeral_username)?;
-    if opts.raw {
-        resolved.render_markdown = false;
-    }
-    if opts.no_tool_calls {
-        resolved.core.no_tool_calls = true;
-    }
-    let use_reflection = resolved.core.reflection_enabled;
-
-    let context_dir = chibi.app.context_dir(ctx_name);
-    let _lock = chibi_core::lock::ContextLock::acquire(
-        &context_dir,
-        chibi.app.config.lock_heartbeat_seconds,
-    )?;
-
-    let mut options = PromptOptions::new(
-        opts.verbose,
-        use_reflection,
-        opts.debug,
-        opts.force_markdown,
-    );
-    if let Some(fb) = fallback {
-        options = options.with_fallback(fb);
-    }
-
-    let md_config = if resolved.render_markdown && !opts.raw {
-        Some(md_config_from_resolved(
-            &resolved,
-            chibi.home_dir(),
-            opts.force_markdown,
-        ))
-    } else {
-        None
-    };
-
-    let mut sink = CliResponseSink::new(
-        output,
-        md_config,
-        opts.verbose,
-        opts.show_tool_calls,
-        opts.show_thinking || resolved.show_thinking,
-    );
-    chibi
-        .send_prompt_streaming(ctx_name, prompt, &resolved.core, &options, &mut sink)
-        .await
-}
-
-fn inspect_context(
-    chibi: &Chibi,
-    context_name: &str,
-    thing: &Inspectable,
-    resolved_config: Option<&ResolvedConfig>,
-    force_markdown: bool,
-    output: &dyn OutputSink,
-) -> io::Result<()> {
-    // Resolve config if not provided
-    let config_holder;
-    let config = if let Some(cfg) = resolved_config {
-        cfg
-    } else {
-        config_holder = resolve_cli_config(chibi, context_name, None)?;
-        &config_holder
-    };
-
-    match thing {
-        Inspectable::List => {
-            output.emit_result("Inspectable items:");
-            for name in <Inspectable as InspectableExt>::all_names_cli() {
-                output.emit_result(&format!("  {}", name));
-            }
-        }
-        Inspectable::SystemPrompt => {
-            let prompt = chibi.app.load_system_prompt_for(context_name)?;
-            if prompt.is_empty() {
-                output.emit_result("(no system prompt set)");
-            } else {
-                output.emit_result(prompt.trim_end());
-            }
-        }
-        Inspectable::Reflection => {
-            let reflection = chibi.app.load_reflection()?;
-            if reflection.is_empty() {
-                output.emit_result("(no reflection set)");
-            } else {
-                output.emit_result(reflection.trim_end());
-            }
-        }
-        Inspectable::Todos => {
-            let todos = chibi.app.load_todos_for(context_name)?;
-            if todos.is_empty() {
-                output.emit_result("(no todos)");
-            } else {
-                let md_cfg = md_config_from_resolved(config, chibi.home_dir(), force_markdown);
-                render_markdown_output(&todos, md_cfg)?;
-                if !todos.ends_with('\n') {
-                    println!();
-                }
-            }
-        }
-        Inspectable::Goals => {
-            let goals = chibi.app.load_goals_for(context_name)?;
-            if goals.is_empty() {
-                output.emit_result("(no goals)");
-            } else {
-                let md_cfg = md_config_from_resolved(config, chibi.home_dir(), force_markdown);
-                render_markdown_output(&goals, md_cfg)?;
-                if !goals.ends_with('\n') {
-                    println!();
-                }
-            }
-        }
-        Inspectable::Home => {
-            output.emit_result(&chibi.home_dir().display().to_string());
-        }
-        Inspectable::ConfigField(field_path) => match config.get_field(field_path) {
-            Some(value) => output.emit_result(&value.to_string()),
-            None => output.emit_result("(not set)"),
-        },
-    }
-    Ok(())
-}
-
-/// Show log entries for a context
-fn show_log(
-    chibi: &Chibi,
-    context_name: &str,
-    num: isize,
-    verbose: bool,
-    resolved_config: &ResolvedConfig,
-    force_markdown: bool,
-    output: &dyn OutputSink,
-) -> io::Result<()> {
-    let entries = chibi.app.read_jsonl_transcript(context_name)?;
-
-    // Select entries based on num parameter
-    let selected: Vec<_> = if num == 0 {
-        entries.iter().collect()
-    } else if num > 0 {
-        let n = num as usize;
-        entries
-            .iter()
-            .rev()
-            .take(n)
-            .collect::<Vec<_>>()
-            .into_iter()
-            .rev()
-            .collect()
-    } else {
-        let n = (-num) as usize;
-        entries.iter().take(n).collect()
-    };
-
-    for entry in selected {
-        match entry.entry_type.as_str() {
-            ENTRY_TYPE_MESSAGE => {
-                output.emit_result(&format!("[{}]", entry.from.to_uppercase()));
-                let md_cfg =
-                    md_config_from_resolved(resolved_config, chibi.home_dir(), force_markdown);
-                render_markdown_output(&entry.content, md_cfg)?;
-                output.newline();
-            }
-            ENTRY_TYPE_TOOL_CALL => {
-                if verbose {
-                    output.emit_result(&format!("[TOOL CALL: {}]\n{}\n", entry.to, entry.content));
-                } else {
-                    let args_preview = if entry.content.len() > 60 {
-                        format!("{}...", &entry.content[..60])
-                    } else {
-                        entry.content.clone()
-                    };
-                    output.emit_result(&format!("[TOOL: {}] {}", entry.to, args_preview));
-                }
-            }
-            ENTRY_TYPE_TOOL_RESULT => {
-                if verbose {
-                    output.emit_result(&format!(
-                        "[TOOL RESULT: {}]\n{}\n",
-                        entry.from, entry.content
-                    ));
-                } else {
-                    let size = entry.content.len();
-                    let size_str = if size > 1024 {
-                        format!("{:.1}kb", size as f64 / 1024.0)
-                    } else {
-                        format!("{}b", size)
-                    };
-                    output.emit_result(&format!("  -> {}", size_str));
-                }
-            }
-            "compaction" => {
-                if verbose {
-                    output.emit_result(&format!("[COMPACTION]: {}\n", entry.content));
-                }
-            }
-            _ => {
-                if verbose {
-                    output.emit_result(&format!(
-                        "[{}]: {}\n",
-                        entry.entry_type.to_uppercase(),
-                        entry.content
-                    ));
-                }
-            }
-        }
-    }
-    Ok(())
-}
-
-/// Set system prompt for a context
-fn set_prompt_for_context(
-    chibi: &Chibi,
-    context_name: &str,
-    arg: &str,
-    verbose: bool,
-) -> io::Result<()> {
-    let content = if std::path::Path::new(arg).is_file() {
-        std::fs::read_to_string(arg)?
-    } else {
-        arg.to_string()
-    };
-    chibi.app.set_system_prompt_for(context_name, &content)?;
-    if verbose {
-        eprintln!("[System prompt set for context '{}']", context_name);
-    }
-    Ok(())
-}
-
-/// Execute from ChibiInput
+/// Handles CLI-specific concerns (context selection, session, username overrides,
+/// help/version, image cache) then delegates command dispatch to core's
+/// `execute_command()`. Session updates are driven by `CommandEffect`.
 async fn execute_from_input(
     input: ChibiInput,
     chibi: &mut Chibi,
@@ -451,44 +248,14 @@ async fn execute_from_input(
     let show_tool_calls = !input.flags.hide_tool_calls || verbose;
     let show_thinking_flag = input.flags.show_thinking || verbose;
 
-    let send_opts = CliSendOptions {
-        raw: input.raw,
-        no_tool_calls: input.flags.no_tool_calls,
-        debug: &input.flags.debug,
-        verbose,
-        force_markdown,
-        show_tool_calls,
-        show_thinking: show_thinking_flag,
-    };
-
-    // Initialize session (executes OnStart hooks)
-    let _ = chibi.init();
-
-    // Auto-destroy expired contexts
-    let destroyed = chibi.app.auto_destroy_expired_contexts(verbose)?;
-    if !destroyed.is_empty() {
-        chibi.save()?;
-        // If our session points to a destroyed context, reset to default
-        if destroyed.contains(&session.implied_context) {
-            session.implied_context = "default".to_string();
-            session.previous_context = None;
-            session.save(chibi.home_dir())?;
-        }
-        output.diagnostic(
-            &format!("[Auto-destroyed {} expired context(s)]", destroyed.len()),
-            verbose,
-        );
-    }
-
     let mut did_action = false;
 
-    // Handle context selection and determine working_context
-    // - working_context: the context we're actually operating on this invocation
-    // - implied_context: persisted in session.json, what you get when no context is specified
+    // --- CLI-specific: context selection ---
+    // working_context: the context we're actually operating on this invocation
+    // implied_context: persisted in session.json, what you get when no context is specified
     let working_context = match &input.context {
         ContextSelection::Current => session.implied_context.clone(),
         ContextSelection::Ephemeral { name } => {
-            // Ephemeral: use context directly WITHOUT mutating session
             let actual_name = resolve_context_name(chibi, session, name)?;
             chibi.app.ensure_context_dir(&actual_name)?;
             output.diagnostic(
@@ -519,48 +286,18 @@ async fn execute_from_input(
         }
     };
 
-    // Touch the working context
-    let current_ctx = working_context.clone();
-    let debug_destroy_at = input.flags.debug.iter().find_map(|k| match k {
-        DebugKey::DestroyAt(ts) => Some(*ts),
-        _ => None,
-    });
-    let debug_destroy_after = input.flags.debug.iter().find_map(|k| match k {
-        DebugKey::DestroyAfterSecondsInactive(secs) => Some(*secs),
-        _ => None,
-    });
-    // Ensure ContextEntry exists before applying destroy settings (fix for #47 regression)
-    if !chibi
-        .app
-        .state
-        .contexts
-        .iter()
-        .any(|e| e.name == current_ctx)
-    {
-        chibi.app.state.contexts.push(ContextEntry::with_created_at(
-            current_ctx.clone(),
-            now_timestamp(),
-        ));
-    }
-    if chibi.app.touch_context_with_destroy_settings(
-        &current_ctx,
-        debug_destroy_at,
-        debug_destroy_after,
-    )? {
-        chibi.save()?;
-    }
-
-    // Handle username override
-    // Persistent (-u) is saved to local.toml; ephemeral (-U) is used for this invocation only
+    // --- CLI-specific: username override ---
     let ephemeral_username: Option<&str> = match &input.username_override {
         Some(UsernameOverride::Persistent(username)) => {
-            let mut local_config = chibi.app.load_local_config(&current_ctx)?;
+            let mut local_config = chibi.app.load_local_config(&working_context)?;
             local_config.username = Some(username.clone());
-            chibi.app.save_local_config(&current_ctx, &local_config)?;
+            chibi
+                .app
+                .save_local_config(&working_context, &local_config)?;
             output.diagnostic(
                 &format!(
                     "[Username '{}' saved to context '{}']",
-                    username, current_ctx
+                    username, working_context
                 ),
                 verbose,
             );
@@ -571,384 +308,90 @@ async fn execute_from_input(
         None => None,
     };
 
-    // SYNC: chibi-json also dispatches commands — check crates/chibi-json/src/main.rs
+    // --- CLI-specific: intercept ShowHelp / ShowVersion ---
     match &input.command {
         Command::ShowHelp => {
             Cli::print_help();
+            return Ok(());
         }
         Command::ShowVersion => {
             output.emit_result(&format!("chibi {}", env!("CARGO_PKG_VERSION")));
+            return Ok(());
         }
-        Command::ListContexts => {
-            let contexts = chibi.list_contexts();
-            let implied = &session.implied_context;
-            for name in contexts {
-                let context_dir = chibi.app.context_dir(&name);
-                let status = chibi_core::lock::ContextLock::get_status(
-                    &context_dir,
-                    chibi.app.config.lock_heartbeat_seconds,
-                );
-                let status_str = status.map(|s| format!(" {}", s)).unwrap_or_default();
-                if &name == implied {
-                    output.emit_result(&format!("* {}{}", name, status_str));
-                } else {
-                    output.emit_result(&format!("  {}{}", name, status_str));
-                }
-            }
-            did_action = true;
-        }
-        Command::ListCurrentContext => {
-            let context_name = &working_context;
-            let context = chibi.app.get_or_create_context(context_name)?;
-            let context_dir = chibi.app.context_dir(context_name);
-            let status = chibi_core::lock::ContextLock::get_status(
-                &context_dir,
-                chibi.app.config.lock_heartbeat_seconds,
-            );
-            let status_str = status.map(|s| format!(" {}", s)).unwrap_or_default();
-            output.emit_result(&format!("Context: {}{}", context_name, status_str));
-            output.emit_result(&format!("Messages: {}", context.messages.len()));
-            if !context.summary.is_empty() {
-                output.emit_result(&format!(
-                    "Summary: {}",
-                    context.summary.lines().next().unwrap_or("")
-                ));
-            }
-            did_action = true;
-        }
-        Command::DestroyContext { name } => {
-            let ctx_name = match name {
-                Some(n) => resolve_context_name(chibi, session, n)?,
-                None => working_context.clone(),
-            };
+        _ => {}
+    }
 
-            if !chibi.app.context_dir(&ctx_name).exists() {
-                output.emit_result(&format!("Context '{}' not found", ctx_name));
-            } else if !output.confirm(&format!("Destroy context '{}'?", ctx_name)) {
-                output.emit_result("Aborted");
-            } else {
-                // Handle session fallback if destroying current context
-                if session
-                    .handle_context_destroyed(&ctx_name, |name| {
-                        chibi.app.context_dir(name).exists()
-                    })
-                    .is_some()
-                {
-                    session.save(chibi.home_dir())?;
-                }
-                chibi.app.destroy_context(&ctx_name)?;
-                output.emit_result(&format!("Destroyed context: {}", ctx_name));
+    // --- resolve CLI-specific context names in command fields ---
+    let command = resolve_command_names(&input.command, chibi, session)?;
+
+    // --- resolve config and build CLI response sink ---
+    let cli_config = resolve_cli_config(chibi, &working_context, ephemeral_username)?;
+
+    let md_config = if cli_config.render_markdown && !input.raw {
+        Some(md_config_from_resolved(
+            &cli_config,
+            chibi.home_dir(),
+            force_markdown,
+        ))
+    } else {
+        None
+    };
+
+    let mut sink = CliResponseSink::new(
+        output,
+        md_config,
+        verbose,
+        show_tool_calls,
+        show_thinking_flag || cli_config.show_thinking,
+    );
+
+    // --- delegate to core ---
+    let effect = chibi_core::execute_command(
+        chibi,
+        &working_context,
+        &command,
+        &input.flags,
+        &cli_config.core,
+        ephemeral_username,
+        output,
+        &mut sink,
+    )
+    .await?;
+    if !matches!(command, Command::NoOp) {
+        did_action = true;
+    }
+
+    // --- CLI-specific: handle CommandEffect for session updates ---
+    match &effect {
+        CommandEffect::ContextDestroyed(name) => {
+            if session
+                .handle_context_destroyed(name, |n| chibi.app.context_dir(n).exists())
+                .is_some()
+            {
+                session.save(chibi.home_dir())?;
             }
-            did_action = true;
         }
-        Command::ArchiveHistory { name } => {
-            let ctx_name = match name {
-                Some(n) => resolve_context_name(chibi, session, n)?,
-                None => working_context.clone(),
-            };
-            if name.is_none() {
-                // Use wrapper with hooks for current context
-                chibi.clear_context(&ctx_name)?;
-            } else {
-                // For named contexts, just clear without hooks (hooks are for interactive use)
-                chibi.app.clear_context(&ctx_name)?;
-            }
-            output.emit_result(&format!(
-                "Context '{}' archived (history saved to transcript)",
-                ctx_name
-            ));
-            did_action = true;
-        }
-        Command::CompactContext { name } => {
-            if let Some(ctx_name) = name {
-                let resolved_name = resolve_context_name(chibi, session, ctx_name)?;
-                api::compact_context_by_name(&chibi.app, &resolved_name, verbose).await?;
-                output.emit_result(&format!("Context '{}' compacted", ctx_name));
-            } else {
-                let resolved = chibi.resolve_config(&working_context, None)?;
-                api::compact_context_with_llm_manual(
-                    &chibi.app,
-                    &working_context,
-                    &resolved,
-                    verbose,
-                )
-                .await?;
-            }
-            did_action = true;
-        }
-        Command::RenameContext { old, new } => {
-            let old_name = match old {
-                Some(n) => resolve_context_name(chibi, session, n)?,
-                None => working_context.clone(),
-            };
-            chibi.app.rename_context(&old_name, new)?;
-            // Update session if we renamed the implied context
-            if session.implied_context == old_name {
+        CommandEffect::ContextRenamed { old, new } => {
+            if session.implied_context == *old {
                 session.implied_context = new.clone();
                 session.save(chibi.home_dir())?;
             }
-            // Also update previous_context if needed
-            if session.previous_context.as_deref() == Some(&old_name) {
+            if session.previous_context.as_deref() == Some(old.as_str()) {
                 session.previous_context = Some(new.clone());
                 session.save(chibi.home_dir())?;
             }
-            output.emit_result(&format!("Renamed context '{}' to '{}'", old_name, new));
-            did_action = true;
         }
-        Command::ShowLog { context, count } => {
-            let ctx_name = match context {
-                Some(n) => resolve_context_name(chibi, session, n)?,
-                None => working_context.clone(),
-            };
-            let config = resolve_cli_config(chibi, &ctx_name, None)?;
-            show_log(
-                chibi,
-                &ctx_name,
-                *count,
-                verbose,
-                &config,
-                force_markdown,
-                output,
-            )?;
-            did_action = true;
-        }
-        Command::Inspect { context, thing } => {
-            let ctx_name = match context {
-                Some(n) => resolve_context_name(chibi, session, n)?,
-                None => working_context.clone(),
-            };
-            // Pass ephemeral username so -U works with -n
-            let config = resolve_cli_config(chibi, &ctx_name, ephemeral_username)?;
-            inspect_context(
-                chibi,
-                &ctx_name,
-                thing,
-                Some(&config),
-                force_markdown,
-                output,
-            )?;
-            did_action = true;
-        }
-        Command::SetSystemPrompt { context, prompt } => {
-            let ctx_name = match context {
-                Some(n) => resolve_context_name(chibi, session, n)?,
-                None => working_context.clone(),
-            };
-            set_prompt_for_context(chibi, &ctx_name, prompt, verbose)?;
-            did_action = true;
-        }
-        Command::RunPlugin { name, args } => {
-            let tool = tools::find_tool(&chibi.tools, name).ok_or_else(|| {
-                io::Error::new(ErrorKind::NotFound, format!("Plugin '{}' not found", name))
-            })?;
-            let args_json = serde_json::json!({ "args": args });
-            let result = tools::execute_tool(tool, &args_json, verbose)?;
-            output.emit_result(&result);
-            did_action = true;
-        }
-        Command::CallTool { name, args } => {
-            let args_str = args.join(" ");
-            let args_json: serde_json::Value = if args_str.is_empty() {
-                serde_json::json!({})
-            } else {
-                serde_json::from_str(&args_str).map_err(|e| {
-                    io::Error::new(
-                        ErrorKind::InvalidInput,
-                        format!("Invalid JSON arguments: {}", e),
-                    )
-                })?
-            };
-
-            let result = chibi
-                .execute_tool(&working_context, name, args_json.clone())
-                .await?;
-
-            if input.flags.force_call_agent {
-                // Tool-first with continuation to LLM
-                let tool_context = format!(
-                    "[User initiated tool call: {}]\n[Arguments: {}]\n[Result: {}]",
-                    name, args_json, result
-                );
-                let fallback = chibi_core::tools::HandoffTarget::Agent {
-                    prompt: String::new(),
-                };
-                send_with_cli_sink(
-                    chibi,
-                    &working_context,
-                    &tool_context,
-                    ephemeral_username,
-                    &send_opts,
-                    Some(fallback),
-                    output,
-                )
-                .await?;
-            } else {
-                // Tool-first with immediate return (default for -P)
-                output.emit_result(&result);
-            }
-            did_action = true;
-        }
-        Command::ClearCache { name } => {
-            let ctx_name = match name {
-                Some(n) => resolve_context_name(chibi, session, n)?,
-                None => working_context.clone(),
-            };
-            chibi.app.clear_tool_cache(&ctx_name)?;
-            output.emit_result(&format!("Cleared tool cache for context '{}'", ctx_name));
-            did_action = true;
-        }
-        Command::CleanupCache => {
-            let resolved = chibi.resolve_config(&working_context, None)?;
-            let removed = chibi
-                .app
-                .cleanup_all_tool_caches(resolved.tool_cache_max_age_days)?;
-            output.emit_result(&format!(
-                "Removed {} old cache entries (older than {} days)",
-                removed,
-                resolved.tool_cache_max_age_days + 1
-            ));
-            did_action = true;
-        }
-        Command::SendPrompt { prompt } => {
-            // Ensure context exists
-            let ctx_name = working_context.clone();
-            if !chibi.app.context_dir(&ctx_name).exists() {
-                let new_context = Context::new(ctx_name.clone());
-                chibi.app.save_and_register_context(&new_context)?;
-            }
-
-            send_with_cli_sink(
-                chibi,
-                &ctx_name,
-                prompt,
-                ephemeral_username,
-                &send_opts,
-                None,
-                output,
-            )
-            .await?;
-            did_action = true;
-        }
-        Command::CheckInbox { context } => {
-            let ctx_name = resolve_context_name(chibi, session, context)?;
-
-            // Peek to see if there are messages
-            let messages = chibi.app.peek_inbox(&ctx_name)?;
-            if messages.is_empty() {
-                output.diagnostic(
-                    &format!("[No messages in inbox for '{}'.]", ctx_name),
-                    verbose,
-                );
-            } else {
-                output.diagnostic(
-                    &format!(
-                        "[Processing {} message(s) from inbox for '{}']",
-                        messages.len(),
-                        ctx_name
-                    ),
-                    verbose,
-                );
-
-                // Ensure context exists
-                if !chibi.app.context_dir(&ctx_name).exists() {
-                    let new_context = Context::new(ctx_name.clone());
-                    chibi.app.save_and_register_context(&new_context)?;
-                }
-
-                send_with_cli_sink(
-                    chibi,
-                    &ctx_name,
-                    chibi_core::INBOX_CHECK_PROMPT,
-                    None,
-                    &send_opts,
-                    None,
-                    output,
-                )
-                .await?;
-            }
-            did_action = true;
-        }
-        Command::CheckAllInboxes => {
-            let contexts = chibi.app.list_contexts();
-            let mut processed_count = 0;
-
-            for ctx_name in contexts {
-                // Peek to see if there are messages
-                let messages = chibi.app.peek_inbox(&ctx_name)?;
-                if messages.is_empty() {
-                    continue;
-                }
-
-                output.diagnostic(
-                    &format!(
-                        "[Processing {} message(s) from inbox for '{}']",
-                        messages.len(),
-                        ctx_name
-                    ),
-                    verbose,
-                );
-
-                send_with_cli_sink(
-                    chibi,
-                    &ctx_name,
-                    chibi_core::INBOX_CHECK_PROMPT,
-                    None,
-                    &send_opts,
-                    None,
-                    output,
-                )
-                .await?;
-
-                processed_count += 1;
-            }
-
-            if processed_count == 0 {
-                output.diagnostic("[No messages in any inbox.]", verbose);
-            } else {
-                output.diagnostic(
-                    &format!("[Processed inboxes for {} context(s).]", processed_count),
-                    verbose,
-                );
-            }
-            did_action = true;
-        }
-        Command::ModelMetadata { model, full } => {
-            let resolved = chibi.resolve_config(&working_context, None)?;
-            let gateway = chibi_core::gateway::build_gateway(&resolved)?;
-            let metadata = chibi_core::model_info::fetch_metadata(&gateway, model).await?;
-            output.emit_result(
-                chibi_core::model_info::format_model_toml(&metadata, *full).trim_end(),
-            );
-            did_action = true;
-        }
-        Command::NoOp => {
-            // No operation - just context switch, already handled above
-        }
+        CommandEffect::None => {}
     }
 
-    // Shutdown session (executes OnEnd hooks)
-    let _ = chibi.shutdown();
-
-    // Automatic cache cleanup
-    let resolved = chibi.resolve_config(&working_context, None)?;
-    if resolved.auto_cleanup_cache {
-        let removed = chibi
-            .app
-            .cleanup_all_tool_caches(resolved.tool_cache_max_age_days)?;
-        if removed > 0 {
-            output.diagnostic(
-                &format!(
-                    "[Auto-cleanup: removed {} old cache entries (older than {} days)]",
-                    removed,
-                    resolved.tool_cache_max_age_days + 1
-                ),
-                verbose,
-            );
-        }
+    // --- CLI-specific: check if auto-destroy removed our session context ---
+    if !chibi.app.context_dir(&session.implied_context).exists() {
+        session.implied_context = "default".to_string();
+        session.previous_context = None;
+        session.save(chibi.home_dir())?;
     }
 
-    // Image cache cleanup
-    let cli_config = resolve_cli_config(chibi, &working_context, None)?;
+    // --- CLI-specific: image cache cleanup ---
     if cli_config.image.cache_enabled {
         let image_cache_dir = chibi.home_dir().join("image_cache");
         match image_cache::cleanup_image_cache(
@@ -971,8 +414,8 @@ async fn execute_from_input(
         }
     }
 
-    // Check for no action and no prompt
-    if !did_action && matches!(input.command, Command::NoOp) {
+    // --- no-action check ---
+    if !did_action && matches!(command, Command::NoOp) {
         return Err(io::Error::new(
             ErrorKind::InvalidInput,
             "No operation specified",
